Update README.md
 /etc/init.d  is a sym link to /etc/rc.d/init.d

Trivial change, but I created a tar that extracted from root to drop all the files in the right place. Doing so replaced the link on my instance with an actual directory that contained the single docker command.  (And that was bad.)  The clarification may save someone else some time.